Correct GetPortTypeId, various cleanup.
git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@2546 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ windows/JackWinNamedPipe.cpp
@@ -35,7 +35,7 @@ int JackWinNamedPipe::Read(void* data, int len)
 {
     DWORD read;
     BOOL res = ReadFile(fNamedPipe, data, len, &read, NULL);
-    if (read != len) {
+    if (read != (DWORD)len) {
         jack_error("Cannot read named pipe err = %ld", GetLastError());
         return -1;
     } else {
@@ -47,7 +47,7 @@ int JackWinNamedPipe::Write(void* data, int len)
 {
     DWORD written;
     BOOL res = WriteFile(fNamedPipe, data, len, &written, NULL);
-    if (written != len) {
+    if (written != (DWORD)len) {
         jack_error("Cannot write named pipe err = %ld", GetLastError());
         return -1;
     } else {
